[Improvement-14331][common] Remove the duplicate ParameterUtils
