+++ dolphinscheduler-task-plugin/dolphinscheduler-task-sagemaker/src/main/java/org/apache/dolphinscheduler/plugin/task/sagemaker/SagemakerTask.java
@@ -29,8 +29,7 @@ import org.apache.dolphinscheduler.plugin.task.api.TaskConstants;
 import org.apache.dolphinscheduler.plugin.task.api.TaskException;
 import org.apache.dolphinscheduler.plugin.task.api.TaskExecutionContext;
 import org.apache.dolphinscheduler.plugin.task.api.model.Property;
-import org.apache.dolphinscheduler.plugin.task.api.parser.ParamUtils;
-import org.apache.dolphinscheduler.plugin.task.api.parser.ParameterUtils;
+import org.apache.dolphinscheduler.plugin.task.api.utils.ParameterUtils;
 
 import org.apache.commons.lang3.StringUtils;
 
@@ -167,7 +166,7 @@ public class SagemakerTask extends AbstractRemoteTask {
 
     private String parseRequstJson(String requestJson) {
         Map<String, Property> paramsMap = taskRequest.getPrepareParamsMap();
-        return ParameterUtils.convertParameterPlaceholders(requestJson, ParamUtils.convert(paramsMap));
+        return ParameterUtils.convertParameterPlaceholders(requestJson, ParameterUtils.convert(paramsMap));
     }
 
     protected AmazonSageMaker createClient() {
